fix "Viewing.Extension.ScreenShotManager.js"
I copied "Viewing.Extension.ScreenShotManager.js". When using this, I found 2 small bugs. 

1: 
"height" and "width" in state[id] are used to represent the height and width of the screenshot. But when this extension's panel becomes undocked, "height" and "width" will also be used to represent the panel's height and width. 

If I want to resize the undocked ScreenShotManager panel, it will soon become full-screen and can not be resize to a smaller one, because the panel's height and width become equal to the viewer container's height and width.

So I renamed the "height" and "width" of the screenshot to "screenshotHeight" and "screenshotWidth".  After renaming, the ScreenShotManager panel could be resized.

2:
In <ContentEditable>, I changed "html={state.screenshotHeight}" into "html={`${state.screenshotHeight}`}". It is because that some problems will occur if the type of this html prop is Number rather than String. The "replace" function of String is used in <ContentEditable> but Number does not has "replace" function. There will be some errors if the html prop is Number.

Also, if the type of html prop is Number, a warning will show in console:
    Warning: Failed prop type: Invalid prop `html` of type `number` supplied to `ContentEditable`, expected `string`.
    in ContentEditable (created by ViewerConfigurator)
    in ViewerConfigurator (created by DatabaseView)
    ......

This seems no longer a problem if using new version of "react-contenteditable" (<ContentEditable>). But now the ScreenShotManager in https://forge-rcdb.autodesk.io/ does not work because of that. So I think it is better to mannually change the type into String.

In https://forge-rcdb.autodesk.io/, if I use ScreenShotManager, what the console shows is:
Uncaught (in promise) TypeError: e.replace is not a function
    at i (configurator.531122013159909c62f5.min.js:1)
    at t.shouldComponentUpdate (configurator.531122013159909c62f5.min.js:1)
......