modify "onResize" in "Viewer.Configurator.js"
I copied "Viewer.Configurator.js" and "Viewing.Extension.ScreenShotManager.js". Sometimes, the height and width of the screenshot did not change after the viewer container resized. So I did some modifications in "Viewer.Configurator.js".

================================================

1:
Before modification, as long as the screenshotManager panel was undocked, the onResize function of this extension would not be triggered even though the window or viewer container resized. It was because that Configurator only called the onResize functions of the "this.state.renderExtension". And for those extensions in "this.state.viewerPanels", their onResize function would not be called. 

But I think these onResize should also be called. For example, when the viewer container resizes, even if the screenshotManager is undocked, it still need to use onResize function to track the height and width of the viewer container. 

So I add some codes to trigger every "viewerPanel.renderable.onResize" when the viewer container resizes. And now the screenshotManager can show and utilize correct height and width of the viewer container when the screenshotManager panel is undocked.

================================================

2:
Before modification, the animation of the "pushRenderExtension" or "popRenderExtension" would change the height and width of the viewer container. However, the animation would not trigger the onResize function of Configurator. 

So I add "this.onResize()" after the "pushRenderExtension" or "popRenderExtension". In this way, the onResize function can be called after the animation ended.   

I did some tries and decided to add "this.onResize()" after pushing or poping an extension. If "this.onResize()" is added directly after the animation, it will have no effects because at that moment the state has not been changed already and there are no extensions in  "this.state.renderExtension". 

With these modifications, the screenshotManager can show and utilize the correct height and width of the viewer container after the animation ends.